+++ src/library/interfaces/collection.d.ts
@@ -6,7 +6,7 @@ export interface ICollection extends Array<IResource> {
     $length: number;
     $toArray: Array<IResource>;
     $is_loading: boolean;
-    $source: 'new' | 'memory' | 'store' | 'server' | 'httpstorage';
+    $source: 'new' | 'memory' | 'store' | 'server';
     $cache_last_update: number;
     data: Array<IDataResource>; // this need disapear is for datacollection
     page: IPage;