+++ src/library/interfaces/params-collection.d.ts
@@ -7,4 +7,5 @@ interface IParamsCollection extends IParams {
     smartfilter?: object;
     page?: IPage;
     storage_ttl?: number;
+    cachehash?: string;  // solution for when we have different resources with a same id
 }